+++ tests/test.py
@@ -1,0 +1,5 @@
+from xulbux import Console
+
+
+Console.log_box_filled("Hello, world!", "Bye, world!")
+Console.log_box_bordered("Hello, world!", "Bye, world!")